api/requirements.txt
Django==2.2.3
psycopg2-binary==2.8.3
djangorestframework==3.9.4
django-filter==2.1.0
Markdown==3.1.1
boto3==1.9.192
django-storages==1.7.1
django-cors-headers==3.0.2
Pillow==6.1.0
